An attempt to detect file changes that occur between build invocations, when there is no significant delay between the invocations. For example, this often happens when testing or benchmarking a build, where the test or benchmark makes changes to the source or build files between the build invocations. This particular problem has caused grief to pretty much every single person who has ever written a functional test for Gradle or a plugin.
More precisely, this change attempts to detect file changes that happened immediately after the completion of the previous build. The approach may be over eager in re-hashing of the final outputs of the previous build, and may need some tweaking, for example, by applying the strategy only to input files.